refactor(logging): restrict level_to_mcp to pub(crate)
The function has exactly one caller within the crate and is not part
of the public API. Move test coverage to an inline cfg(test) module
and remove the external integration test import.
